validate database and fix chat

+++ backend/src/main/java/com/qm/bookstore/qm_bookstore/dto/user/request/UserGetAllRequest.java
@@ -13,5 +13,5 @@ import lombok.experimental.FieldDefaults;
 public class UserGetAllRequest extends BaseGetAllRequest {
     String username;
     String email;
-    Long roleId;
+    Integer roleId;
 }
